refactor(web): only export react components in jsx files (contexts/snackbar)
