+++ web/src/contexts/snackbar/index.jsx
@@ -1,0 +1,11 @@
+import { createContext } from "react";
+
+
+export const SnackbarContext = createContext({
+    snackbar: {
+        open: false,
+        severity: "info",
+        message: "",
+    },
+    setSnackbar: () => { },
+});
